feat(api): add route deactivate admin member
