+++ api/tests/unit/application/admin-members/index_test.js
@@ -5,6 +5,23 @@ const adminMembersRouter = require('../../../../lib/application/admin-members');
 const securityPreHandlers = require('../../../../lib/application/security-pre-handlers');
 
 describe('Unit | Application | Router | admin-members-router', function () {
+  describe('GET /api/admin/admin-members/me', function () {
+    it('should return a response with an HTTP status code 200', async function () {
+      // given
+      const adminMember = domainBuilder.buildAdminMember();
+      sinon.stub(adminMemberController, 'getCurrentAdminMember').returns(adminMember);
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(adminMembersRouter);
+
+      // when
+      const { statusCode } = await httpTestServer.request('GET', '/api/admin/admin-members/me');
+
+      // then
+      expect(adminMemberController.getCurrentAdminMember).to.have.be.called;
+      expect(statusCode).to.equal(200);
+    });
+  });
+
   describe('GET /api/admin/admin-members', function () {
     it('should return a response with an HTTP status code 200 when user has role "SUPER_ADMIN"', async function () {
       // given
@@ -18,17 +35,17 @@ describe('Unit | Application | Router | admin-members-router', function () {
       const { statusCode } = await httpTestServer.request('GET', '/api/admin/admin-members');
 
       // then
+      expect(securityPreHandlers.checkUserHasRoleSuperAdmin).to.have.be.called;
+      expect(adminMemberController.findAll).to.have.be.called;
       expect(statusCode).to.equal(200);
-      sinon.assert.calledOnce(adminMemberController.findAll);
     });
 
     it('should return a response with an HTTP status code 403 if user does not have the rights', async function () {
       // given
-      const adminMembers = [domainBuilder.buildAdminMember()];
-      sinon.stub(adminMemberController, 'findAll').returns(adminMembers);
       sinon
         .stub(securityPreHandlers, 'checkUserHasRoleSuperAdmin')
         .callsFake((request, h) => h.response().code(403).takeover());
+
       const httpTestServer = new HttpTestServer();
       await httpTestServer.register(adminMembersRouter);
 
@@ -36,6 +53,7 @@ describe('Unit | Application | Router | admin-members-router', function () {
       const { statusCode } = await httpTestServer.request('GET', '/api/admin/admin-members');
 
       // then
+      expect(securityPreHandlers.checkUserHasRoleSuperAdmin).to.have.be.called;
       expect(statusCode).to.equal(403);
     });
   });
@@ -177,22 +195,40 @@ describe('Unit | Application | Router | admin-members-router', function () {
   });
 
   describe('PATCH /api/admin/admin-members/{id}', function () {
-    it('should return a response with an HTTP status code 200 when user has role "SUPER_ADMIN"', async function () {
-      // given
-      const updatedAdminMember = domainBuilder.buildAdminMember();
-      sinon.stub(adminMemberController, 'updateAdminMember').returns(updatedAdminMember);
-      sinon.stub(securityPreHandlers, 'checkUserHasRoleSuperAdmin').returns(true);
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(adminMembersRouter);
+    describe('when user has role "SUPER_ADMIN"', function () {
+      it('should return a response with an HTTP status code 200', async function () {
+        // given
+        const updatedAdminMember = domainBuilder.buildAdminMember();
+        sinon.stub(adminMemberController, 'updateAdminMember').returns(updatedAdminMember);
+        sinon.stub(securityPreHandlers, 'checkUserHasRoleSuperAdmin').returns(true);
+        const httpTestServer = new HttpTestServer();
+        await httpTestServer.register(adminMembersRouter);
 
-      // when
-      const { statusCode } = await httpTestServer.request('PATCH', '/api/admin/admin-members/1', {
-        data: { type: 'admin-members', attributes: { role: ROLES.SUPER_ADMIN } },
+        // when
+        const { statusCode } = await httpTestServer.request('PATCH', '/api/admin/admin-members/1', {
+          data: { type: 'admin-members', attributes: { role: ROLES.SUPER_ADMIN } },
+        });
+
+        // then
+        expect(securityPreHandlers.checkUserHasRoleSuperAdmin).to.have.be.called;
+        expect(statusCode).to.equal(200);
       });
 
-      // then
-      expect(statusCode).to.equal(200);
-      sinon.assert.calledOnce(adminMemberController.updateAdminMember);
+      it('should return 400 if the role value to update is invalid', async function () {
+        // given
+        sinon.stub(securityPreHandlers, 'checkUserHasRoleSuperAdmin').returns(true);
+        const httpTestServer = new HttpTestServer();
+        await httpTestServer.register(adminMembersRouter);
+
+        // when
+        const { statusCode } = await httpTestServer.request('PATCH', '/api/admin/admin-members/1', {
+          data: { type: 'admin-members', attributes: { role: 'INVALID_ROLE' } },
+        });
+
+        // then
+        expect(securityPreHandlers.checkUserHasRoleSuperAdmin).not.to.have.be.called;
+        expect(statusCode).to.equal(400);
+      });
     });
 
     it('should return a response with an HTTP status code 403 if user does not have the rights', async function () {
@@ -210,8 +246,64 @@ describe('Unit | Application | Router | admin-members-router', function () {
       });
 
       // then
+      expect(securityPreHandlers.checkUserHasRoleSuperAdmin).to.have.be.called;
       expect(statusCode).to.equal(403);
-      sinon.assert.notCalled(adminMemberController.updateAdminMember);
+    });
+  });
+
+  describe('PUT /api/admin/admin-members/{id}/deactivate', function () {
+    const method = 'PUT';
+    const url = '/api/admin/admin-members/1/deactivate';
+
+    describe('When user has role "SUPER_ADMIN"', function () {
+      it('should return a response with an HTTP status code 200', async function () {
+        // given
+        sinon
+          .stub(adminMemberController, 'deactivateAdminMember')
+          .callsFake((request, h) => h.response('ok').code(204));
+        sinon.stub(securityPreHandlers, 'checkUserHasRoleSuperAdmin').returns(true);
+        const httpTestServer = new HttpTestServer();
+        await httpTestServer.register(adminMembersRouter);
+
+        // when
+        const { statusCode } = await httpTestServer.request(method, url);
+
+        // then
+        expect(securityPreHandlers.checkUserHasRoleSuperAdmin).to.have.be.called;
+        expect(adminMemberController.deactivateAdminMember).to.have.be.called;
+        expect(statusCode).to.equal(204);
+      });
+
+      it('should return a 400 error when ID is invalid', async function () {
+        // given
+        sinon.stub(securityPreHandlers, 'checkUserHasRoleSuperAdmin').returns(true);
+        const httpTestServer = new HttpTestServer();
+        await httpTestServer.register(adminMembersRouter);
+
+        // when
+        const { statusCode } = await httpTestServer.request(method, '/api/admin/admin-members/invalidID/deactivate');
+
+        // then
+        expect(securityPreHandlers.checkUserHasRoleSuperAdmin).not.to.have.be.called;
+        expect(statusCode).to.equal(400);
+      });
+    });
+
+    it('should return a response with an HTTP status code 403 if user does not have the rights', async function () {
+      // given
+      sinon.stub(adminMemberController, 'updateAdminMember').returns('ok');
+      sinon
+        .stub(securityPreHandlers, 'checkUserHasRoleSuperAdmin')
+        .callsFake((request, h) => h.response().code(403).takeover());
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(adminMembersRouter);
+
+      // when
+      const { statusCode } = await httpTestServer.request(method, url);
+
+      // then
+      expect(securityPreHandlers.checkUserHasRoleSuperAdmin).to.have.be.called;
+      expect(statusCode).to.equal(403);
     });
   });
 });
